acpi: reintroduce LNKS
Commit 4540409 (acpi: LNKS is not needed, 2012-08-07) removed LNKS because
it bascally worked by chance: _CRS returns something else then one of the
possible resources from _PRS, _DIS would not really disable the interrupt,
and there was no _SRS method.  It just happened to work because Linux
special-cases the SCI all over the place.

Unfortunately, the code after the patch is also against the spec, and
it breaks FreeBSD because it treats IRQ 9 polarity as active low without
the Interrupt() entry.  Numeric _PRT entries are handled the same in
Linux and FreeBSD, as active-low, and under Linux it just happens to
trigger another special casing of SCI which sets SCI up based on the
override in the MADT, ignoring the DSDT completely.

This patch adds back the LNKS, but without using the PIIX register for
LNKA in its methods.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>