boot: move menu to F11, reserve F12 for network boot
The SMBIOS specification requests that F12 is used for network boot,
so add that shortcut and move the existing menu to F11 (similar to
what is used on some bare-metal machines).

The implementation of F12 will prioritize entries with a network
device PNP type over all the other items of the boot list.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>